Re-Structured code
